+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - BlockFi parser: added new export format.
 - BlockFi parser: added "Crypto Transfer" and "BIA Withdraw" transaction types.
 - Conversion tool: added parser for Binance.US exchange.
+- Celsius parser: added "Swap in" and "Swap out" transaction types.
 ### Changed
 - Conversion tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))
 - Accounting tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))